# docs/api/openapi/v1/all.json
{
  "components": {
    "schemas": {
      "ChangePasswordRequest": {
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateBranchRequest": {
        "properties": {
          "addressLine": {
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityRootId": {
            "format": "uuid",
            "type": "string"
          },
          "governorate": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "entityRootId",
          "name",
          "addressLine",
          "city",
          "governorate",
          "phone"
        ],
        "type": "object"
      },
      "CreateEntityRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "registrationNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "typeCode": {
            "type": "string"
          },
          "workspaceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "workspaceId",
          "name",
          "typeCode",
          "registrationNumber"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SetupRequest": {
        "properties": {
          "locale": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "workspaceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "workspaceSlug": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
      "UpdateBranchRequest": {
        "properties": {
          "addressLine": {
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "type": [
              "null",
              "string"
            ]
          },
          "governorate": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "addressLine",
          "city",
          "governorate",
          "phone"
        ],
        "type": "object"
      },
      "UpdateEntityRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "registrationNumber": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "name",
          "registrationNumber"
        ],
        "type": "object"
      },
...
      "UseRecoveryCodeRequest": {
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "recoveryCode": {
            "type": "string"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Aggregate of all Balsm modules \u2014 generated. Do not hand-edit.",
    "title": "Balsm API (aggregate)",
    "version": "v1"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/admin/auth/change-password": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "AdminAuth"
        ]
      }
    },
    "/api/v1/admin/auth/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "AdminAuth"
        ]
      }
    },
...
    "/api/v1/admin/auth/recovery/use": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UseRecoveryCodeRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UseRecoveryCodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UseRecoveryCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "AdminAuth"
        ]
      }
    },
    "/api/v1/admin/auth/setup": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "AdminAuth"
        ]
      }
    },
...
    "/api/v1/admin/branches": {
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Branches"
        ]
      }
    },
    "/api/v1/admin/branches/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Branches"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Branches"
        ]
      }
    },
    "/api/v1/admin/branches/{id}/reactivate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Branches"
        ]
      }
    },
...
    "/api/v1/admin/entities": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "workspaceId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeInactive",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      }
    },
    "/api/v1/admin/entities/{entityId}/branches": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "entityId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeInactive",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Branches"
        ]
      }
    },
    "/api/v1/admin/entities/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      }
    },
    "/api/v1/admin/entities/{id}/reactivate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "Entities"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5550/"
    }
  ],
...
}